+++ src/Cattech.Optimizer.Pro.Infrastructure/Hardware/SensorProvider/IHardwareMonitorAbstraction.cs
@@ -30,6 +30,7 @@ internal enum InternalSensorType
     Current,
     Power,
     TimeSpan,
+    Timing,
     Other
 }